requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
pyyaml>=6.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

